% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression Class.}
\description{
\code{ridgereg()} calculates a number of statistics (see section \code{"Fields"}) by default given both arguments: formula, data and lambda. 
It also stores diverse functions implemented using RC object oriented programming. Find more information regarding these functions on section \code{"Methods"}.
}
\section{Fields}{

\describe{
\item{\code{formula}}{an object of class "formula".}

\item{\code{data}}{a data frame.}

\item{\code{regression_coefficients}}{the regression coefficients of the model calculated using the QR decomposition.}

\item{\code{fitted_values}}{the fitted values.}

\item{\code{lambda}}{a integer.}

\item{\code{arguments}}{a vector containing the function call.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the coefficients.}

\item{\code{initialize(formula, data, lambda)}}{Initialize function calculates all values needed from formula, data and lambda.}

\item{\code{predict(newdata)}}{Calculate predictions based on fitted values and new x data.}

\item{\code{print()}}{Same as show, but allows for ridgereg$print(). Will break print(ridgereg) that show allows.}
}}

\examples{
library(dplyr)
library(mlbench)
data("BostonHousing")
set.seed(42)

dataSet <- dplyr::select(BostonHousing, -c(chas))

example <- ridgereg$new(formula = medv~., data = dataSet, lambda = 10**6)
example$regression_coefficients

}
\references{
Reference Classes: \url{http://adv-r.had.co.nz/R5.html}


Vignettes: \url{http://r-pkgs.had.co.nz/vignettes.html}
}
