% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup.R
\name{ridgereg_model}
\alias{ridgereg_model}
\title{Function for generating caret custom model.}
\usage{
ridgereg_model()
}
\value{
\code{ridgereg_model} returns a custom model object for the caret package.
}
\description{
\code{ridgereg_model} generates a custom model for use with the caret package.
}
\examples{
rr <- ridgereg_model()

}
\references{
\url{https://topepo.github.io/caret/using-your-own-model-in-train.html}
}
