% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Function for generating plot of mean arrival delays at different airports.}
\usage{
visualize_airport_delays()
}
\value{
\code{visualize_airport_delays} returns a plot.
}
\description{
\code{visualize_airport_delays} generates a map plot with color scale for mean arrival delay at different airports.Dataset used is the nycflights13.
}
\examples{
visualize_airport_delays()

}
\references{
\url{https://www.rdocumentation.org/packages/nycflights13/versions/1.0.1}
}
